Variable names for Global and Brazil have been checked to eliminate ' and define acronyms.  "Variables with types.xls" contains the result.
Proposed name changes would improve consistency in acronym use.
Capitalization consistency is not proposed, because we plan automated
fixes to that.  Categories column suggests possible categories, but this
is incomplete and probably not needed.  Capitalization column indicates
variables where capitalization inconsistencies were noted; these will be
corrected automatically.  Standard is:
1. Stocks are capitalized (every word)
2. Acronyms are capitalized, except units.
3. Camel case is to be converted to separations with _

List of abbreviations needs to be checked by Steve. If their use is NOT
indicated with a "NO" in the Use? column, then proposed names in the
other sheet have been consistently modified to eliminate this
abbreviation.