adds definitions to vbls
Detail on new tabs.
SP regional Vbls contains a list of regional variables, consistent with
BRAZIL from tab 1.  It provides documentation for each.
SP Global VBLS contains a list of vbls in the current Global module,
along with documentation.
SP Changes to vbls contains a list of deletions and additions to vbls in
BRAZIL and Global in the first tab..  I have not deleted/added vbls to
other regions in the list in the first tab.

Model changes are in SVN build 3906.